requirements.txt
tiktoken==0.4.0 
backoff==2.2.1 
openai==0.28.0 
azure.cognitiveservices.speech==1.32.1
RPi.GPIO==0.7.1
luma.core==2.4.1
luma.lcd==2.11.0
Pillow==9.4.0
nicegui==1.3.13
pyalsaaudio==0.9.2
python-dotenv==1.0.0